Basically starting from scratch? #39
Oh boy, here we go. So I learned that the way I SHOULD have done this is by making the png relative, and the rest of everything else absolute. So I now know that I have to change the position type and position of every single thing I've been working on for the past who knows long.